Added PowerShell support for on demand code generation from Package Manager Console
