+++ WebApiProxy.Tasks/WebApiProxy.Tasks.csproj
@@ -55,6 +55,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Infrastructure\CSharpGenerator.cs" />
     <Compile Include="Models\Configuration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ProxyGenerationTask.cs" />
@@ -69,9 +70,16 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <None Include="install.ps1" />
+    <None Include="init.ps1">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Include="install.ps1">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Include="packages.config" />
-    <None Include="uninstall.ps1" />
+    <None Include="uninstall.ps1">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Include="WebApiProxy.config">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       <SubType>Designer</SubType>
@@ -87,7 +95,9 @@
       <Generator>TextTemplatingFilePreprocessor</Generator>
       <LastGenOutput>CSharpProxyTemplate.cs</LastGenOutput>
     </None>
-    <None Include="webapiproxy.snippet" />
+    <None Include="webapiproxy.snippet">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
   <ItemGroup>
     <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
@@ -97,6 +107,11 @@
       <Project>{94508206-8537-4088-923b-e7fcb0539903}</Project>
       <Name>WebApiProxy.Core</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="WebApiProxyCSharp.psm1">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
