+++ WebApiProxy.Tasks/WebApiProxyCSharp.psm1
@@ -1,0 +1,32 @@
+﻿function WebApiProxy-Generate-CSharp() {
+	$project = Get-Project
+    
+    $projectPath = [System.IO.Path]::GetDirectoryName($project.FullName)
+
+	$root = (Join-Path $projectPath "WebApiProxy\")
+
+	$rootSpaces = "$root"
+
+	$config = [WebApiProxy.Tasks.Models.Configuration]::Load($rootSpaces);
+
+    $generator = New-Object WebApiProxy.Tasks.Infrastructure.CSharpGenerator -ArgumentList @($config)
+    
+    
+    
+    $fileName = (Join-Path $projectPath "WebApiProxy\WebApiProxy.generated.cs")
+    
+	Write-Host "Generating proxy code..."
+
+    $source = $generator.Generate()
+    
+    $result = New-Item $fileName `
+          -ItemType "file" -Force `
+          -Value $source
+    
+    $item = $project.ProjectItems.AddFromFile($fileName)
+
+	Write-Host "Done."
+
+}
+
+Export-ModuleMember "WebApiProxy-Generate-CSharp"
