+++ WebApiProxy.Tasks/init.ps1
@@ -1,0 +1,4 @@
+﻿param($installPath, $toolsPath, $package)
+
+Add-Type -Path (Join-Path $installPath "build\WebApiProxy.Tasks.dll")
+Import-Module (Join-Path $toolsPath "WebApiProxyCSharp.psm1") -DisableNameChecking
